Still can't edit

+++ src/main/java/service/UserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface UserService {
 
     void saveUser(User user);
-    void deleteUserById(int id);
+    void deleteUserById(Long id);
     void editUser(User user);
     List<User> getAllUsers();
-    User getUserById(int id);
+    User getUserById(Long id);
 }
